Adds links and images in html
creates a html file and uses the <a> tag to link webpages and then uses the <img> tag to link images.

This part helped me to understand the usage of <a> and <img> tag and further more gave me a clear understanding of html attributes and tags.

The attributes used are:
- src
-alt
-href